GearUp: Integrated latest from dev.
[git-p4: depot-paths = "//depot/traktor/": change = 10411]

+++ code/Net/Replication/ReplicatorTypes.h
@@ -8,7 +8,7 @@ namespace traktor
 	namespace net
 	{
 
-typedef uint32_t handle_t;
+typedef uint64_t handle_t;
 
 	}
 }
